Refactoring: moved generic brain to the common package.
