+++ src/main/java/com/leokom/games/chess/player/legal/brain/common/Brain.java
@@ -2,9 +2,10 @@ package com.leokom.games.chess.player.legal.brain.common;
 
 import com.leokom.games.chess.engine.Move;
 import com.leokom.games.chess.engine.Position;
+import com.leokom.games.commons.brain.GenericBrain;
 
 /**
  * Chess-specific brain
  */
-public interface Brain extends GenericBrain<Position, Move>{
+public interface Brain extends GenericBrain<Position, Move> {
 }
